refactor: move BatchSearch root to a sub page
