+++ src/pages/BatchSearch.vue
@@ -1,75 +1,9 @@
 <template>
-  <div class="batch-search container h-100 pt-4">
-    <v-wait class="batch-search__wait" for="load haveBatchSearch">
-      <fa slot="waiting" class="d-flex mx-auto mt-5" icon="circle-notch" size="2x" spin />
-      <template v-if="hasBatchSearch">
-        <div class="d-flex flex-wrap align-items-center">
-          <batch-search-filter-query class="batch-search__search-bar my-1" />
-          <batch-search-clear-filters
-            class="batch-search__clear-filter-btn m-1"
-            route-name="batch-search"
-            :local-search-params="LOCAL_SEARCH_PARAMS"
-          />
-        </div>
-        <batch-search-table />
-      </template>
-      <template v-else>
-        <div class="batch-search__none text-center">
-          <div class="batch-search__none__message b-table-empty-row" v-html="noBatchSearch" />
-        </div>
-      </template>
-    </v-wait>
-  </div>
+  <router-view></router-view>
 </template>
 
 <script>
-import { mapGetters } from 'vuex'
-
-import utils from '@/mixins/utils'
-import BatchSearchTable from '@/components/BatchSearchTable'
-import BatchSearchClearFilters from '@/components/BatchSearchClearFilters'
-import BatchSearchFilterQuery from '@/components/BatchSearchFilterQuery'
-const LOCAL_SEARCH_PARAMS = Object.freeze({
-  query: true,
-  publishState: false,
-  project: false,
-  dateStart: true,
-  dateEnd: true,
-  state: true,
-  order: true,
-  sort: true
-})
 export default {
-  name: 'BatchSearches',
-  components: {
-    BatchSearchFilterQuery,
-    BatchSearchClearFilters,
-    BatchSearchTable
-  },
-  mixins: [utils],
-  data() {
-    return { LOCAL_SEARCH_PARAMS }
-  },
-  computed: {
-    ...mapGetters('batchSearch', ['hasBatchSearch']),
-    howToLink() {
-      return '#/docs/all-batch-search-documents'
-    },
-    noBatchSearch() {
-      return this.$t('batchSearch.empty', { howToLink: this.howToLink })
-    }
-  },
-  mounted() {
-    if (!this.hasBatchSearch) {
-      return this.getBatchSearches()
-    }
-  },
-  methods: {
-    async getBatchSearches() {
-      this.$wait.start('load haveBatchSearch')
-      await this.$store.dispatch('batchSearch/getBatchSearches', { init: true })
-      this.$wait.end('load haveBatchSearch')
-    }
-  }
+  name: 'BatchSearch'
 }
 </script>
