+++ src/pages/BatchSearchList.vue
@@ -5,7 +5,11 @@
       <template v-if="hasBatchSearch">
         <div class="d-flex flex-wrap align-items-center">
           <batch-search-filter-query class="batch-search-list__search-bar my-1" />
-          <batch-search-clear-filters class="batch-search-list__clear-filter-btn m-1" />
+          <batch-search-clear-filters
+            class="batch-search-list__clear-filter-btn m-1"
+            route-name="batch-search"
+            :local-search-params="LOCAL_SEARCH_PARAMS"
+          />
         </div>
         <batch-search-table />
       </template>
@@ -26,6 +30,17 @@ import BatchSearchTable from '@/components/BatchSearchTable'
 import BatchSearchClearFilters from '@/components/BatchSearchClearFilters'
 import BatchSearchFilterQuery from '@/components/BatchSearchFilterQuery'
 
+const LOCAL_SEARCH_PARAMS = Object.freeze({
+  query: true,
+  publishState: false,
+  project: false,
+  dateStart: true,
+  dateEnd: true,
+  state: true,
+  order: true,
+  sort: true
+})
+
 export default {
   name: 'BatchSearchList',
   components: {
@@ -34,6 +49,9 @@ export default {
     BatchSearchTable
   },
   mixins: [utils],
+  data() {
+    return { LOCAL_SEARCH_PARAMS }
+  },
   computed: {
     ...mapGetters('batchSearch', ['hasBatchSearch']),
     howToLink() {
@@ -57,6 +75,7 @@ export default {
   }
 }
 </script>
+
 <style lang="scss" scoped>
 .batch-search-list__none__message {
   padding: 0.75em;
